Use Accept-Language header to set language
This is based on the instructions here, where there's the example function:
`extract_locale_from_accept_language_header`:
http://guides.rubyonrails.org/i18n.html#setup-the-rails-application-for-internationalization

> In practice, more robust code is necessary to do this reliably. Iain
> Hecker's http_accept_language library or Ryan Tomayko's locale Rack
> middleware provide solutions to this problem.

So now I can update the language in my browser preferences, and forttree
will display "latest posts" in the language I choose, if it's japanese,
spanish or danish. Still more work to do for internationalizing all the
text on the site.